Added functionality for movement of a sprite / labels for login
Created a change to game window once user  connects to the server.
Added moving sprite of jeffs head to the game window. Was working on sending movement to the server. Need to think about how to create a sprite once a player joins and also a good way to send updated coordinates of people moving so their board gets updated.